OpenVDB: Make filepath of current frame available
This makes `abs_path` in the OpenVDB modifier return the path to the
frame that should currently be read (taking into account the current
frame, the frame offset, and frame overrides).

This is a bit hackish, as it duplicates the frame calculation and path
generation code, but it was the quickest and easiest way to achieve
this, in contrast to adapting and exposing the existing functions to
work with this use-case.