Cycles: Several emitters per leaf in light tree
This commits makes it possible for leaf nodes in the
light tree to have more than one emitter. This reduces
the maximum depth of the light tree which makes the
traversal of the tree faster but the sampling code
is slower if a leaf with several emitters is found.

Technical details:
-- If cost of splitting in build is larger than the
   energy of the node then we create a leaf instead
   of splitting.
-- Maximum emitters per leaf is now set to 64
-- During traversal if a leaf with several emitters
   is found then we sample one of these based on
   the importance of each emitter. This is calculated
   on the fly and therefore a CDF generation without
   dynamic allocation was implemented.
-  Moved the distribution id to node id lookup from
   light_distribution_pdf() to light_bvh_pdf().