Cycles: Remove split kernel
This changes removes split kernel implementation and majority
of related boiler plate code on device level.

The goal is to free up some room for the new split kernel
implementation.

The OpenCL becomes unusable with this change. CPU, CUDA, and
OptiX are still working. CPU still passes regression tests.

There are some more possibilities to simplify code, like remove
the requested features and CPUKernelFunctions abstraction, but
it can happen as a followup development.