Cycles: Boiler plate code for the new split kernel functions
The implementation of kernels places in the kernel/integrator folder to
make some distinction between new code being written and the existing.

Currently, all the kernels have a single KernelGlobals argument and an
empty implementation. It would be nice if there is a way to have a
minimal and standard arguments, as it simplifies things in the future.

The new "section" of code is marked as such. Not sure if it's helping
or not, but kind of feels it makes it easier to see functions which
are to be removed or, at least, re-evaluated.

The way how kernels are defined and declared follows the legacy macro,
which I'm not huge fan of, but don't feel like looking into more clear
ways of replacing them is a good time investment at this time. It is
all localized, and doesn't matter too much for the actual code of the
kernels implementation.