Cycles: Add boiler plate for PathTrace class
As per design, the idea of this class it to implement the kernel graph
and all the logic needed for adaptive path tracing of a big tile on the
given (multi)device.

This commit is very minimalistic, could be seen as one big TODO, but it
feels better to put all files in place early on.

The change introduces the new folder called integrator in the root of
Cycles sources. The reasoning includes the following thoughts:

- There will be more auxiliary classes such as CPU side of a queue,
  perhaps variance calculation, adaptive stopping criteria and many
  things like that.

- For the long time it felt that we need more separation in Cycles code
  structure to separate scene state and rendering algorithms.