Cycles: Pull make_unique into the ccl namespace
Allows to allocate object instances which are guarded by unique_ptr
in a more modern fashion.

This is a C++14 feature, which is higher than typically used C++11
in the Cycles code. Since Blender is C++17 it makes sense to move
to a newer standard in Cycles as well. All this is applicable if
there are no other places in Cycles which are relying on a more
modern C++ already.