Cycles: Boiler plate code for kernel queues
The goal of this change is lay down an initial version of API which
will allow path tracer to queue kernels for execution.

This is done by introducing new `DeviceQueue` class which is a higher
level abstraction over concepts like OpenCL queue or CUDA stream. Its
purpose is to take overhead of kernels invocation. The queue is created
by a device, and after that it becomes self-sufficient.

The queue is implemented as a "stub" for CPU device.

This change also contains changes which is basically an initial pass
over render graph implementation.

In order to have all bits tested had to rake over `Session::render()`
function so now render result is always empty. The old code is ifdefed
so that it is more visible which parts of code are to be brought back
in a proper design.

Next steps will include implementation of some resource management so
that integrator kernels can be actually invoked.