Cycles: Boiler plate for actual kernels invocation
Now the integrator kernels will be invoked by the path traces.

For now the path state is hard-coded to 1. Will be done properly
as one of the nearest future changes.

Another thing which needs to be changed in the nearest future is
the render buffer management. We really need to communicate it
to the BlenderSession to start seeing actual pixels!