Cycles: Properly initialize camera rays
The idea is to pass KernelWorkTile to the generate_camera_rays
kernel so that it can have access to screen-space coordinates.
Some fields of the KernelWorkTile are unused and are either to
be removed in the future, or a simplified version of a struct
is to be used.

The device side of changes is related on making it to queue is
aware of where in the bug tile if is operating on. Currently
done as an API of DeviceQueue, and with time might become more
clear whether such work tile concept should only be done for
path tracing related queue.

The path tracer change is rather annoying and a bit messy: it
evaluates the render graph for every pixel in a progressive
refine manner. It is slow and single threaded, and code is not
folded very flexible yet. The good thing is that it allows to
make it so kernels to output proper data.

The output kernels needs an adjustment, so that they write to
an offset within the render buffer. Only modified the surface
kernel, because otherwise other kernels will override the value
and make it so nothing is visible.