Refactor: Move Cycles devices to sub-folders
The idea is to have per-device type folder, where more than one file is
allowed to be existing. This is something what was already happening
for CUDA and OpenCL devices. Do the same for CPU and OptiX. Also move
the device implementation files to the folder.

The way how OpenCL and OptiX class hierarchy is implemented needs a bit
of refactor to streamline it. It will happen when we will be adding
support of CUDA and OptiX back.