Refactor: Move Cycles device declarations closer to implementation
Avoid one header where functions for all devices are declared. Also,
made it so caller will not care about compiled-in features: it should
be possible to request device initialization and enumeration, and if
the compiled-in support is missing it will act similar to case when
the system does not have the requested device.

Makes it so things are more properly black-boxed and separated.
There is still some possible combing to be done, but think this change
is a good first iteration.