Cycles: Initial pass of adding work scheduling for path tracer
Touches a lot of little things which will be needed for a final version
of work scheduler, putting API in place.

The scheduling logic is still CPU-centric, but is now hidden behind a
more manageable API.

The biggest thing to figure out is the big tile render buffer memory
and per-device "work" memory.