Cycles: Simplify render buffers management
Avoid per-integrator-queue render buffers, which also avoids copy
of render results from one buffer to another.

Now on CPU all integrator queues are sharing same render buffer
created for the big tile and are writing to different parts of
this buffer.

For a single-GPU device there will be one integrator queue and one
render buffer on the device side.

Multi-device would need to have some more tricky buffer management
that it originally looked like to me; data transfers are to be
avoided as much as possible. So ideally it would be some per-device
statically allocated buffers, with possibly some re-arrangement
during rendering to balance things better. From the memory ownership
point of view it will mean per-device render buffers and multiple
integrator queues for CPU. So from this point of view the idea of
path tracing context becomes obsolete.

Remove the PathTraceContext. Maybe later it can become PathTraceWork
with an implementation of "wavefront management". But for now lets
keep things more localized to PathTrace for easier refactoring.