Cycles: Make path tracer aware of start sample of tile manager
This is a part of "bring viewport rendering back" project.
This is something what is required to have with the current way of
how viewport drawing is done. In the future it might make sense to
have PathTrace persistent in the Session, so that it can keep track
of sample and resolution divider.

However, this new API will likely be used to achieve the "add more
samples to existing render" feature.