Cycles: Initial support of viewport rendering
Rather straightforward idea: request path trace to copy its render
buffers to the given display buffer. There is some more glue logic
to pass samples, resolution divider, and proper offset/stride to
the kernel.

For some reason viewport is not that interactive: seems that there
is no redraw on lower resolution results. Could be because of the
locking nature of the `render_samples()` call, or could be something
else. Forcing resolution divider in the tile manager, or setting
pixel size to 8x clearly shows proper behavior, so it's something
with the updates rather than an issue in implementation of this
part of the code.

The TileManager is still the source of truth for current sample
and resolution. This will be moved out in the further development.