Cycles: Remove DeviceWorkTile
It is really the same as KernelWorkTile, especially after we remove the
buffer pointer (hopefully, in the future).

There is no need to have a duplicate entity, just use KernelWorkTIle in
the API. Simplifies some of initialization when passing work tile to an
actual kernels.

Opens possibility to schedule multiple samples to GPU.

Had to introduce min(uint, uint) and solve some introduced deduction
issues by mix of int and uint in the existing code.