Cycles: Bring progress report back
Is a bit weird mixture of Progress object and a callback.

The idea is that Progress object is used to communicate progress within
the current big tile from PathTrace to owner of the progress object.

The callback is called after that, so that the (in this case) Session
can assemble full status and report it to the interface.

On a positive side the cancel callback becomes unneeded.