Cycles: move integrator specific logic out of device queue
The CUDA integrator queue was mostly CUDA independent and highly coupled to
PathTraceWorkTiled. Seems better to move the tiled integrator logic into a
single class, and keep the queue implementations simple to implement for each
device type.

The particular motivation for this change was scheduling of most queued
kernels, which was getting even more complicated.