Cycles: Make viewport navigation more interactive
Achieved by making render session be resettable more often than the
configured timeout. More detailed explanation of logic can be found
in the comment in `ready_to_reset()`.

Makes it so camera navigation in classroom scene gives feedback as
soon as possible.

The next logical trick on this topic would be to calculate resolution
divider based on the rendering performance, so that we can show most
detailed picture while navigating.