Cycles: Rename Denoiser to DenoiserPipeline
The goal is to have clarity on entity names.

The proposed/desired convention:

- Device: low-level "glue" logic, which takes care of running
  functions on a particular device. Algorithm implementation
  does not belong to device.

- Denoiser: breaks down specific denoising algorithm into a
  function calls which happens on a device. It does not belong
  to the device because denoising can happen regardless of where
  noisy image came from: could be file, could be single render
  buffer, could be multiple render buffers rendered by multiple
  devices.

- DenoiserPipeline: takes care of performing all steps needed to
  denoise a sequence of frames.

In an ideal world DenoiserPipeline will implement a frame accessor
of some sort, and use Denoiser to perform actual denoising on a
configured device/denoising algorithm.

This change only modifies the class name, making room for implementing
an actual denoising algorithms as dedicated entities. Further cleanup
in naming and structure is needed. It will be revisited once the
cross-frame denoising will be worked on as an official core
functionality.