Cycles: Support denoiser in the new integrator
Implemented as own classes which provide higher level API for rendering
pipeline. Currently placed in the integrator folder, which feels a bit
strange, but is easy to move these files around.

The idea is the following:

- Have every implementation to implement Denoiser interface, which
  goes away from previous implementation on a device level. This allows
  to more naturally consider Device a way to invoke kernels, without
  business logic.

- Currently Denoiser operates on an entire RenderBuffer. In the future
  this would need to be changed to operate on a list of render buffers
  to support denoising of renders done on multi-device. Seems to be
  "just" a straightforward extension of the currently implemented
  simple API.

Currently Session is requesting denoising as a dedicated call of
PathTrace. This is something I am not really sure about: in a way this
feels wrong and Session can/should have direct access to denoiser. But
on another hand, having denoising done in the PathTracer allows to have
a more centralized and comprehensive logic around decision of how many
samples to render before update, when to start denoising and things
like this.

From actual implementation is only OpenImageDenoise with color buffer
is implemented. The rest would need a support from kernel.