Cycles: Fix denoiser not working after recent changes
The change which introduced check of denoising parameters made it
so denoiser in PathTrace was never created: constructor was checking
denosiing parameters with self.

For now use a special function from constructor which ensures
initialization is fully done correctly.

In the coming future the synchronization and locks will be redone,
so while this is a bit annoying change it is only to unlock current
development.