Cycles: refactor CPU/megakernel for more flexible execution order
Previously it would call kernels in a fixed order, which becomes harder with
a more complex state machine for SSS and volumes. Further, there was code
duplication in that kernels would be indicate the next kernel to execute, and
check if the current kernel needs to be executed.

Instead use queued_kernel on both CPU and GPU, and implement the megakernel as
a switch statement using this value.