Cycles: Initial implementation of GPUDisplay
The idea of GPUDisplay is to take care of viewport texture update
and drawing in a way that is the same exact from API point of view
for both CPU and GPU rendering.

This change moves a lot of code and refactors it in a more reusable
and isolated building blocks to implement the GPUDIsplay for Blender
viewport.

The CPU rendering is same exact from users perspective.

The GPU rendering currently does not use CUDA OpenGL buffers, which
can potentially cause degradation in the performance.

There is also a temporary half buffer in BlenderGPUDisplay, because
it is not clear yet how to create OpenGL context, so that we can
update texture from different thread than the drawing.

The update of the GPUDisplay is performed by the Session. Ideally,
this will be done in PathTracer so that it can avoid unnecessary
updates if sample rendered quick enough. This behavior will need a
bit deeper change in a way how TIleManager, Session::render() and
PathTracer are communicating to each other.