Cycles: Make buffer parameters more coupled to PathTraceWork
It is annoying to pass scaled parameters to every call. More correct
approach it seems to be to make them more "sticky": they become a
property of path trace work, together with the render buffers.

The idea is to have the following in the long term:
  - PathTrace allocates render buffers for every PathTraceWork
  - PathTrace updates parameters of every PathTraceWork on resolution
    divider change.
  - PathTrace re-allocates the buffer when balancing is needed to
    accomodate for performance difference of different render devices.

This way PathTraceWork always knows its buffer pointer and effective
parameters and always uses them in all the API calls.