Cycles: Silence unused kg arguments in CPU kernels
The kernel globals might not be used by some of the kernels.
This change makes it so such kernels do not generate a strict compiler
warning.

The motivation is to still be able to see unused arguments (which
could be left-over of refactor and are to be removed) without causing
"polluting" of the compile log with warnings mentioned above.

Using C++17 attribute, since this feels to be the most portable across
compilers solution.