Cycles: Move GPUDisplay to PathTrace
The end goal is to have PathTrace to have full control over things like
resolution divider, number of samples scheduled for rendering, picking
the best moment to perform denoising based on an empirical measurements
in order to achieve best responsive interface as well as the best
render times.

This change moves ownership of the GPUDisplay from Session to PAthTrace
so that PathTrace can start using it directly.

Should be no functional changes so far, just preparing code and API
needed for the end goal.