Cycles: Move samples scheduling logic to own class
The idea is to consolidate all logic which is related on samples and
denoising scheduling to a single class, which has an overview of how
expensive each task is.

So far should be no functional changes. Still preparation for moving
resolution divider and viewport samples number logic away from the
TileManager.