Cycles: Fix missing denoise start sample update
It is being updated separately from parameters, even though is stored
in the parameters.

Solved by moving an update of denoiser parameters to the initialization
of the new render iteration. The modification check is now done in the
PathTrace.

Good thing is that now it is possible to remove buffer lock, meaning
interface will never be frozen on denoiser changes. Additionally,
rendering now should never get out of sync between passes and denoiser
configuration.