Cycles: Move resolution divider to RenderScheduler
Had to move RenderScheduler away from PathTrace into Session.
This way it is simpler to perform scene update in a threading-safe
manner without making PathTrace API more complicated.

So now the flow goes as following:
- Session requests work from the RenderScheduler
- Session updates scene for the possibly changed resolution
- PathTrace renders the given work, which might involve path tracing,
  denoising, and display update.
- PathTrace informs the scheduler how long specific task took, so that
  the scheduler can adjust its scheduling policy.

The user level change is that the start resolution is now ignored.
It will be replaced with a more automated code based on the timing.

There is still some work to be done to support multiple big tiles.

Still not entirely happy, but think we need to let the matter to rest
a bit and re-iterate when we will be adding big tiles support.