Cycles: Schedule multiple samples to viewport render
The idea is to schedule 1 sample at the lower sample ranges, giving
a quick update on a noisy image, but schedule more samples and do less
often updates at higher number of samples.

There is a magic code which decides how often to do updates, which
might need some further tweaking.

Simple startup file and BMW27 on GPU are about 4% faster in viewport
now.

It is important to note that navigation is always instant. This means
that the longer render times between updates does not have affect on
how instant navigation can start.