Cycles: Remove film conversion to byte buffer
The motivation is that such conversion is faster to be performed as
a fragment shader.

For platforms and configurations which can not use fragment shader
an alternative would be to implement it in the GPUDisplay level.

If this functionality is really something what a lot of Cycles
adopters are needing, we can reconsider this. Until then is better
to keep an API small and clean.