Cycles: Remove dead code from Session
The code is not reachable, not testable, and unlikely to be needed for
the final implementation. Remove it to make ongoing development and
maintenance cleaner.

For the reference we can always consult code from the previous Cycles
releases.

There are still some `#if 0` in the `session.cpp` but they are not in
a way and can be removed later on.