Cycles: Remove or comment out end-of-life code
The goal is to make API clean, and stop worrying about keeping it
compilable. The most interesting bits of implementation which we
will need to reference for the new implementation are `#if 0`-ed.

The code which is either migrated or which is known to be not needed
is removed.

Summary:

- Remove `DeviceTask` and `Device::task_foo`
  It is replaced with either queue or CPUKernels.

- Comment out NLM denoiser.
  Can probably remove it. But lets have a quick meeting about it.

- Remove old OpenImageDenoiser implementation.
  It is now implemented in integrator/denoiser.

- Commented out render/denoiser. It is to be ported to the new
  denoiser implementation in inetgrator/. Also, if NLM is removed then
  feature-parity is not really possibly at this time.

- Comment out device_denoiser.