Cycles: Move adaptive sampling convergence test to own kernel
This change makes it so the check is only done after all contribution
is known at the sample. It also makes it so path tracing kernel is
faster because it doesn't do per-bounce convergence test.

Makes it so that CPU samples distribution works similar to the
master branch.

The timing on CPU is way closer to the master branch, but more accurate
measurement with proper compilation flags is needed.

GPU adaptive sampling is temporarily broken.