Cycles: Scale all passes with adaptive samples count
Do it in the RenderBuffers, similar to how combined pass was handled.

While the scaling with samples count is slower than doing it on the
device, the scaling with samples count is still to be performed.

In practice the downside is that per-pixel inverse sample count is now
done from a single thread. However, this can be improved by using tbb
to scale the pass from multiple threads.

The benefit of this approach is that the pass scaling kernel is not
needed, which solves ambiguity about when to run it. It also simplifies
code in a sense that the scaling logic is not duplicated in the kernel.