Cycles: Fix dependency cycle between device update and scheduler
Scheduler needs to know adaptive sampling settings, the device
update needs to know scheduled resolution divider, and the old
adaptive sampling code was using fields which are only available
after device update.

This change fixes this dependency.

The downside is that there is no more check for existence of AUX
pass. Assume that the session is in consistent state and that when
adaptive sampling is enabled toe AUX pass exists.