Cycles: Simplify TileManager
Prepare it for being used for big tile scheduling and not worry about
samples and sample ranges.

Rather straightforward removal of unused code. The only part which
still does not feel fully ideal is that the Session now needs to hold
buffer parameters. Not sure whether there is a better place for them.

The pixel progress is now fully disabled. It was not updated anyway,
so within the current state of the project this is not a regression.
The status report needs to be worked on and this is part of the TODO.