Cycles: Some basements for the OptiX denoiser
The actual logic is not implemented, but some initial APIs are in
place.

The main purpose of this change is to make sure the device for OptiX
denoiser is actually created and available.

Still some work is needed to create denoiser on an appropriate device
but this better to happen as a followup development, because it will
require some deeper logic change somewhere.