Cycles: Bring OptiX denoiser back
Basically putting old code to be used by the new denoiser
implementation in the integrator.

The main denoising function in the OptiXDevice was split,
to make it easier to follow and understand what's going on.

The current limitation is that currently it only works when
rendering on the OptiX device. Attempts to render on CPU and
denoise on OptiX will currently error out (error is printed
to the console, denoising is not run). This will be worked
on next.