Cycles: Fix threading issues in the PathTrace render cancel
The actual flow is a bit of a mystery to me, but there were cases
when `PathTrace::cancel()` was waiting for the `PathTrace::render()`
to finish, but the `render_cancel_.is_requested` was cleared so that
the render_pipeline was not quitting.

The new code seems more logical to me, but really wouldn't mind a
second pair of eyes here.