Cycles: Partial fix for the viewport passes
Blender's addon should use same constants as the PassType.

One thing which is still not working in viewport is Emission and
Background passes when there is an object with Glass shader: the
object is still shaded. The final F12 render of the Emission and
background is correct. So not immediately clear what's the issue.