Cycles: Fix transmission pass leaking into Combined
Was happening when there are transmission objects in the scene and the
transmission passes are not enabled. The pass offset in the KernelFilm
were left at 0, making it so `kernel_accum_light` believed the pass is
to be written, and the writing was happening into the combined pass.

Maybe we should initialize all passes offsets to PASS_UNUSED, so that
on an incorrect usage we'll have a memory error instead of silently
corrupting the combined pass?

This change fixes the startup file of myself, which is a glass monkey
on a plane.